so much work done. but basically - reworked the delivery system. now we have customers, they try to switch queues if there's a freer one near them. but only by one cuz they dum dum. request is done, now i need to redo the storage. so, currently all i have left to do for this is create boxes with different materials, sizes and forms. stuff them with some item just for old time's sake. create this box, keep track of it. first request must always be deposit. maybe spawn a dozen of other boxes, just incase. we don't keep track of them though, they're there to finely work as a replacement for the actual request (i.e. we broke original but we have backup)
rework the storage area to be more boring and instead - just check if something entered it and is still. if so, count the order.requestDelivered or whatever true. we still keep track of this item. if any item falls outta arenaBounds we ask the arenaBounds to remove that order from the list.

and now the actual deposit system, where we create a new requestee, who is waiting to pick up the item that was left here. we reset the "order thingy" on the item on the shelf before creating him, and set the order on the item to the requestee's

now if the item is brought to the requestee, they accept it and then  i need to make a more detailed system for this. so  like.  if any of the materials fit it's already 1/4 better than a random item given to him. oh, and once we give the item to the requestee, THAT item is removed. not the one they actually requested, but the one they received.

and also  missions. for the shifts .  i  keepe forgetting that we're making the shifts gamemode before endless.  just.   endless is already working fine-ish :( . . . .           . .
okay so this wall of text idc

добавить счетчики

4 не выполненных заданий за смену - -голова
за каждое невыпол. задание - -очки, прогрессирующие (а-ля оранжевые космонафты)
счётчик 3 - увеличивает вероятность выподания событий при невыполнений заданий заказчиков (несущественно но повышает до конца смены)
случайные события - когад происходит N кол-во событий, события просиходят или нет (если не происходят, увеличивается многократно по кол-ву неудачных непроизошедших событий, если случилось - сбросилось)  + +ование невыполнения заданий (каждый раз, несущественно)

записанными должны быть
(я) важные механики требуемые для обучения игрока для начального игрока
(всей кучей) сюжет туториала

новая механика
3 точки на часах которые короче жизнь. 0 жизней - бабах тобi крантi